Stupid Update
Check previous update for my talk, I just forgot to make the function an int (stupid ahha). thats all im fixing. CHANGES SIZE ALOT THO.
Program Size: Code=1072 RO-data=32 RW-data=14612 ZI-data=2284  with t = 1.15297075

+++ task3.c
@@ -6,7 +6,7 @@ const int DetThr=0x69000;
 int mod_out;
 int i,j,sample[256];
 
-float Timing_Synchronization(int *Samples, int *Coeff) {
+int Timing_Synchronization(int *Samples, int *Coeff) {
 	int accReal = 0;
 	int accImaginary = 0;
 	int limit = 256;
